requirements.txt
Django==5.2.6
djangorestframework==3.15.2
Pillow==10.4.0
python-decouple==3.8
requests==2.31.0
cryptography==41.0.7
coverage==7.11.0
